split config into smaller files
